tagging_pipeline/requirements.txt
psycopg2-binary
python-dotenv
pylint
pytest
openai
requests
nltk
gensim
spacy